+++ test-app/app/serializers/Polymorphic/WorkoutSerializer.ts
@@ -1,4 +1,5 @@
 import { DreamSerializer } from '../../../../src/index.js'
 import Workout from '../../models/Polymorphic/Workout.js'
 
-export const WorkoutSerializer = (workout: Workout) => DreamSerializer(Workout, workout).attribute('name')
+export const WorkoutSerializer = (workout: Workout) =>
+  DreamSerializer(Workout, workout).attribute('name').rendersOne('workoutType')