+++ test-app/db/migrations/1751044835745-create-polymorphic-workout.ts
@@ -5,7 +5,7 @@ export async function up(db: Kysely<any>): Promise<void> {
     .createTable('polymorphic_workouts')
     .addColumn('id', 'bigserial', col => col.primaryKey())
     .addColumn('polymorphic_workout_type_id', 'bigint', col =>
-      col.references('polymorphic_workout_types.id').onDelete('restrict').notNull()
+      col.references('polymorphic_workout_types.id').onDelete('restrict')
     )
     .addColumn('name', 'varchar(255)')
     .addColumn('created_at', 'timestamp', col => col.notNull())